Add files via upload
This excel file contains all of the 2024 rosters for all teams.  Please use this as a guide on what column names need to be there.  Import this file and it will update every player on every team, it's that simple.  If you just want to update the Jets and Giants for example, remove all other teams besides these 2 and the program will update just these 2 teams.